+++ README.md
@@ -1,5 +1,17 @@
-# GTRaM
+# Silent genes
 
-![Java CI with Maven](https://github.com/ielis/GTRaM/workflows/Java%20CI%20with%20Maven/badge.svg)
+![Java CI with Maven](https://github.com/ielis/SilentGenes/workflows/Java%20CI%20with%20Maven/badge.svg)
 
-Genes, TRanscripts, and More
+A library that lets you silently work with genes and transcripts.
+
+The project is currently in *pre-alpha* stage.
+
+## Library structure
+
+*Silent genes* consists of several modules:
+- 
+- `silent-genes-model` defines the common properties of genes and transcripts,
+- `silent-genes-gencode` provides genes and transcripts as specified by [Gencode consortium](https://www.gencodegenes.org/),
+- `silent-genes-jannovar` provides genes and transcripts from the [Jannovar](https://github.com/charite/jannovar) transcript databases,
+- `silent-genes-simple` gives you a handful of real-life genes, mostly useful for unit testing
+
