--- gtram-gencode/src/main/java/xyz/ielis/gtram/gencode/io/GencodeParser.java
@@ -1,23 +1,0 @@
-package xyz.ielis.gtram.gencode.io;
-
-import org.monarchinitiative.svart.GenomicAssembly;
-import xyz.ielis.gtram.gencode.model.GencodeGene;
-
-import java.nio.file.Path;
-
-public class GencodeParser implements Iterable<GencodeGene> {
-
-    private final Path gencodeGtfPath;
-    private final GenomicAssembly genomicAssembly;
-
-    public GencodeParser(Path gencodeGtfPath, GenomicAssembly genomicAssembly) {
-        this.gencodeGtfPath = gencodeGtfPath;
-        this.genomicAssembly = genomicAssembly;
-    }
-
-    @Override
-    public GeneIterator iterator() {
-        return new GeneIterator(gencodeGtfPath, genomicAssembly);
-    }
-
-}
