--- gtram-gencode/src/main/java/xyz/ielis/gtram/gencode/io/GtfFeature.java
@@ -1,13 +1,0 @@
-package xyz.ielis.gtram.gencode.io;
-
-enum GtfFeature {
-
-    GENE,
-    TRANSCRIPT,
-    EXON,
-    CDS,
-    UTR,
-    START_CODON,
-    STOP_CODON,
-    SELENOCYSTEINE
-}
