--- gtram-gencode/src/main/java/xyz/ielis/gtram/gencode/io/GtfFrame.java
@@ -1,8 +1,0 @@
-package xyz.ielis.gtram.gencode.io;
-
-enum GtfFrame {
-    NA,
-    ZERO,
-    ONE,
-    TWO
-}
