--- gtram-gencode/src/main/java/xyz/ielis/gtram/gencode/io/GtfSource.java
@@ -1,6 +1,0 @@
-package xyz.ielis.gtram.gencode.io;
-
-enum GtfSource {
-    ENSEMBL,
-    HAVANA
-}
