--- gtram-gencode/src/main/java/xyz/ielis/gtram/gencode/model/EvidenceLevel.java
@@ -1,7 +1,0 @@
-package xyz.ielis.gtram.gencode.model;
-
-public enum EvidenceLevel {
-    VERIFIED,
-    MANUALLY_ANNOTATED,
-    AUTOMATICALLY_ANNOTATED
-}
