--- gtram-gencode/src/main/java/xyz/ielis/gtram/gencode/model/Status.java
@@ -1,7 +1,0 @@
-package xyz.ielis.gtram.gencode.model;
-
-public enum Status {
-    KNOWN,
-    NOVEL,
-    PUTATIVE
-}
