--- gtram-model/src/main/java/module-info.java
@@ -1,4 +1,0 @@
-module xyz.ielis.gtram.model {
-    requires transitive org.monarchitiative.svart;
-    exports xyz.ielis.gtram.model;
-}
