+++ silent-genes-gencode/src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module xyz.ielis.gtram.gencode {
+    requires transitive xyz.ielis.silent.genes.model;
+    requires org.slf4j;
+
+    exports xyz.ielis.silent.genes.gencode.model;
+    exports xyz.ielis.silent.genes.gencode.io;
+}
