+++ silent-genes-gencode/src/main/java/xyz/ielis/silent/genes/gencode/io/GtfFeature.java
@@ -1,0 +1,12 @@
+package xyz.ielis.silent.genes.gencode.io;
+
+enum GtfFeature {
+    GENE,
+    TRANSCRIPT,
+    EXON,
+    CDS,
+    UTR,
+    START_CODON,
+    STOP_CODON,
+    SELENOCYSTEINE
+}
