+++ silent-genes-gencode/src/main/java/xyz/ielis/silent/genes/gencode/io/GtfFrame.java
@@ -1,0 +1,8 @@
+package xyz.ielis.silent.genes.gencode.io;
+
+enum GtfFrame {
+    NA,
+    ZERO,
+    ONE,
+    TWO
+}
