+++ silent-genes-gencode/src/main/java/xyz/ielis/silent/genes/gencode/io/GtfSource.java
@@ -1,0 +1,6 @@
+package xyz.ielis.silent.genes.gencode.io;
+
+enum GtfSource {
+    ENSEMBL,
+    HAVANA
+}
