+++ silent-genes-gencode/src/main/java/xyz/ielis/silent/genes/gencode/model/EvidenceLevel.java
@@ -1,0 +1,7 @@
+package xyz.ielis.silent.genes.gencode.model;
+
+public enum EvidenceLevel {
+    VERIFIED,
+    MANUALLY_ANNOTATED,
+    AUTOMATICALLY_ANNOTATED
+}
