+++ silent-genes-gencode/src/main/java/xyz/ielis/silent/genes/gencode/model/GencodeTranscript.java
@@ -1,0 +1,18 @@
+package xyz.ielis.silent.genes.gencode.model;
+
+
+import xyz.ielis.silent.genes.model.Transcript;
+
+public interface GencodeTranscript extends Transcript {
+
+    String id();
+
+    String type();
+
+    Status status();
+
+    String name();
+
+    EvidenceLevel evidenceLevel();
+
+}
