+++ silent-genes-gencode/src/main/java/xyz/ielis/silent/genes/gencode/model/Status.java
@@ -1,0 +1,7 @@
+package xyz.ielis.silent.genes.gencode.model;
+
+public enum Status {
+    KNOWN,
+    NOVEL,
+    PUTATIVE
+}
