+++ silent-genes-model/src/main/java/module-info.java
@@ -1,0 +1,5 @@
+module xyz.ielis.silent.genes.model {
+    requires transitive org.monarchitiative.svart;
+
+    exports xyz.ielis.silent.genes.model;
+}
