+++ silent-genes-model/src/main/java/xyz/ielis/silent/genes/model/Gene.java
@@ -1,0 +1,14 @@
+package xyz.ielis.silent.genes.model;
+
+import java.util.stream.Stream;
+
+/**
+ * Gene is a region in a genome that has at least one transcript.
+ */
+public interface Gene extends Located {
+
+    // TODO - add identifiers
+
+    Stream<Transcript> transcripts();
+
+}
