+++ silent-genes-model/src/main/java/xyz/ielis/silent/genes/model/Spliced.java
@@ -1,0 +1,20 @@
+package xyz.ielis.silent.genes.model;
+
+import org.monarchinitiative.svart.Coordinates;
+
+import java.util.List;
+
+/**
+ * The interface to represent entities that consist of exons and introns.
+ * <p>
+ * The implementors must ensure:
+ * <ul>
+ *     <li>exons are arranged in ascending order using the coordinates,</li>
+ *     <li>exons do not overlap</li>
+ * </ul>
+ */
+public interface Spliced {
+
+    List<Coordinates> exons();
+
+}
